extract parse_annotation_result
