+++ lua/annotation-tool/search/fzf_lua.lua
@@ -7,12 +7,14 @@ local function load_deps()
 	local logger = require('annotation-tool.logger')
 	local lsp = require('annotation-tool.lsp')
 	local config = require('annotation-tool.config')
+	local parser = require('annotation-tool.search.parser')
 
 	return {
 		core = core,
 		preview = preview,
 		logger = logger,
 		lsp = lsp,
+		parser = parser,
 		config = config
 	}
 end
@@ -46,177 +48,6 @@ local function safe_truncate_utf8(str, max_chars, suffix)
 	-- 这个函数会确保不在多字节字符中间截断
 	local truncated = vim.fn.strcharpart(str, 0, max_chars - vim.fn.strchars(suffix))
 	return truncated .. suffix
-end
-
----解析 LSP 返回的标注数据，提取并拆分为内容和笔记的条目列表。
----@param result table LSP 返回的标注结果，包含 note_files 和 workspace_path 字段。
----@return table 标注条目列表，每个条目包含内容或笔记的单行文本、完整内容、文件信息及元数据。
-local function parse_annotations_result(result)
-	local deps = load_deps()
-	local annotations = {}
-
-	if not result or not result.note_files or #result.note_files == 0 then
-		return annotations
-	end
-
-	local workspace_path = result.workspace_path
-	local current_file = vim.fn.expand('%:p')
-
-	---将原始标注内容和笔记分割为按行的条目，并生成包含元数据的内容和笔记条目列表。
-	local function create_annotation_entries(og_content, og_note, base_info)
-		local content_entries = {}
-		local note_entries = {}
-
-		-- 处理内容行 - 只有有内容时才创建content条目
-		if og_content and og_content ~= "" then
-			local content_lines = {}
-			for line in og_content:gmatch("[^\r\n]+") do
-				local trimmed = line:gsub("^%s*(.-)%s*$", "%1")
-				if trimmed ~= "" then -- 跳过空行
-					table.insert(content_lines, trimmed)
-				end
-			end
-
-			-- 只有当有有效内容行时才创建条目
-			if #content_lines > 0 then
-				for i, line in ipairs(content_lines) do
-					table.insert(content_entries, {
-						file = base_info.file,
-						content = line, -- 单行内容
-						full_content = og_content, -- 完整内容用于预览
-						full_note = og_note, -- 完整笔记用于预览
-						position = base_info.position,
-						range = base_info.range,
-						note_file = base_info.note_file,
-						workspace_path = base_info.workspace_path,
-						line_info = string.format("内容第%d行", i),
-						is_content_line = true,
-						line_number = i,
-						entry_type = "content"
-					})
-				end
-			end
-		end
-
-		-- 处理笔记行 - 只有有笔记时才创建note条目
-		if og_note and og_note ~= "" then
-			local note_lines = {}
-			for line in og_note:gmatch("[^\r\n]+") do
-				local trimmed = line:gsub("^%s*(.-)%s*$", "%1")
-				if trimmed ~= "" then -- 跳过空行
-					table.insert(note_lines, trimmed)
-				end
-			end
-
-			-- 只有当有有效笔记行时才创建条目
-			if #note_lines > 0 then
-				for i, line in ipairs(note_lines) do
-					table.insert(note_entries, {
-						file = base_info.file,
-						note = line, -- 单行笔记
-						full_content = og_content, -- 完整内容用于预览
-						full_note = og_note, -- 完整笔记用于预览
-						position = base_info.position,
-						range = base_info.range,
-						note_file = base_info.note_file,
-						workspace_path = base_info.workspace_path,
-						line_info = string.format("笔记第%d行", i),
-						is_note_line = true,
-						line_number = i,
-						entry_type = "note"
-					})
-				end
-			end
-		end
-
-		return content_entries, note_entries
-	end
-
-	for _, note_file_info in ipairs(result.note_files) do
-		local note_file = note_file_info.note_file
-
-		-- 获取标注内容
-		local file_path = workspace_path .. "/.annotation/notes/" .. note_file
-
-		-- 使用 pcall 进行错误处理
-		local ok, file_content = pcall(vim.fn.readfile, file_path)
-		if not ok then
-			deps.logger.warn("无法读取文件: " .. file_path)
-			goto continue
-		end
-
-		-- 输出调试信息
-		deps.logger.debug("尝试读取文件: " .. file_path)
-		deps.logger.debug_obj("文件内容", file_content)
-
-		-- 提取标注内容和笔记
-		local content = ""
-		local note = ""
-		local in_notes_section = false
-		local in_selected_text_section = false
-		local in_code_block = false
-		local in_frontmatter = false
-		local position = { line = 0, character = 0 }
-		local range = { start = { line = 0, character = 0 }, ["end"] = { line = 0, character = 0 } }
-
-		for _, line in ipairs(file_content) do
-			-- 处理 frontmatter
-			if line:match("^%-%-%-") then
-				in_frontmatter = not in_frontmatter
-			elseif in_frontmatter then
-				-- 跳过 frontmatter 内容
-				goto inner_continue
-			elseif line:match("^## Selected Text") then
-				in_selected_text_section = true
-				in_notes_section = false
-			elseif line:match("^## Notes") then
-				in_notes_section = true
-				in_selected_text_section = false
-				in_code_block = false
-			elseif in_selected_text_section then
-				-- 在 Selected Text 部分
-				if line:match("^```") then
-					in_code_block = not in_code_block
-				elseif in_code_block then
-					-- 提取代码块内的内容，保持原始格式
-					if content ~= "" then
-						content = content .. "\n"
-					end
-					content = content .. line
-				end
-			elseif in_notes_section then
-				-- 在 Notes 部分
-				if note ~= "" then
-					note = note .. "\n"
-				end
-				note = note .. line
-			end
-			::inner_continue::
-		end
-
-		-- 使用新的拆分逻辑
-		local base_info = {
-			file = current_file,
-			position = position,
-			range = range,
-			note_file = note_file,
-			workspace_path = workspace_path
-		}
-
-		local content_entries, note_entries = create_annotation_entries(content, note, base_info)
-
-		-- 将content和note条目都添加到annotations中，但标记类型
-		for _, entry in ipairs(content_entries) do
-			table.insert(annotations, entry)
-		end
-		for _, entry in ipairs(note_entries) do
-			table.insert(annotations, entry)
-		end
-
-		::continue::
-	end
-
-	return annotations
 end
 
 ---过滤标注条目根据搜索模式
@@ -342,7 +173,7 @@ function M.search_annotations(options)
 	deps.logger.debug_obj("第一个标注", options.annotations_result.note_files[1])
 
 	-- 解析标注数据
-	local annotations = parse_annotations_result(options.annotations_result)
+	local annotations = deps.parser.parse_annotations_result(options.annotations_result)
 
 	if #annotations == 0 then
 		deps.logger.info("解析后无有效标注")
@@ -456,7 +287,7 @@ function M.search_annotations(options)
 							end
 
 							-- 更新全局annotations变量
-							annotations = parse_annotations_result(new_result)
+							annotations = deps.parser.parse_annotations_result(new_result)
 							deps.logger.info("标注删除成功，列表已刷新")
 
 							-- 重新启动搜索
